reorganize files inside a new App folder
